ref-os-iot: Updates to XDK daemon support
It was found that XDK works OK, but is missing the avahi
and other support modules. This prevents the XDK IDE from
identifying the device.

ref-os-iot has included xdk-damon and supported modules undee
intel-xdk package feature. This is an issue because in 5xx,
xdk-daemon is not included as part of image where as it is
in edison.

Spliting the package feature definition to include daemon
as a separate item.

Signed-off-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Paasto, Matti <matti.paasto@intel.com>
Signed-off-by: Jukka Laitinen <jukka.laitinen@intel.com>