Add only one beignet library to the target
Currently there are two versions of the beignet library being
built. This adds only 570x for 550x/570x boards, and drops out
minnowmax one. For edison this uses beignet-none, which will
cause an error as there is no beignet for edison.

This also leaves out the ocl-icd. It can be easily enabled by adding
it to the FEATURE_PACKAGES_opencl.

Signed-off-by: Jukka Laitinen <jukka.laitinen@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>