ref-os-iot/meta-edison: Fix for BT not working.
The latest refkit has introduced a common-tools package
group via commit

commit fc1d6c2cfe57ec2d33dffc87e6f2e40c582e6898
Author: Simo Kuusela <simo.kuusela@intel.com>
Date:   Wed Feb 8 13:49:46 2017 +0200

This introduced connman-client to full-image. In Edison
connman is no more used as it would need a heavily patched
systemd for it to work with wpa-supplicant and to use the
existing tools available. Thus connman was removed from
ref-os-iot images for Edison.

This patch removes connman again from full-image for
Edison.

Signed-off-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Katila, Tuomas <tuomas.katila@intel.com>