Fix kernel configuration warnings from iwlwifi
Disabling iwlwifi from kernel configuration causes a lot of irrelevant
warnings, because some dependent configs are still attempted in the kernel meta.

As we don't really disable the iwlwifi but just replace the modules from another recipe,
we don't want to disable the configs. Instead, just delete the in-tree built modules
after compilation and package the ones from the backport-iwlwifi recipe instead.

Signed-off-by: Jukka Laitinen <jukka.laitinen@intel.com>
Reviewed-by: Katila, Tuomas <tuomas.katila@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>