meta-ref-os-iot/meta-edison: Introduce sketch autostart service
Currently Edison doesn't start the sketch downloaded from using
Arduino IDE upon reboot. This is a feature needed when Edison is
to be used in a final real world product.

This patch introduces the needed scripts and service which can
be used when deploying a final product.

The scripts are not part of the default image.

Custom Image needs to include the recipe in the image configuration
as bellow,

In file: meta-ref-os-iot/meta-edison/recipes-core/images/full-image.bb

 IMAGE_INSTALL_append = "\
     ${MACHINE_EXTRA_RRECOMMENDS} \
     autostart-sketch \
 "

Signed-off-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Laitinen, Jukka <jukka.laitinen@intel.com>
Reviewed-by: Kalugin, Evgeny <evgeny.kalugin@intel.com>