Enable wcove sleep clock support for audio
This implements a clock provider driver for wcove 32kHz clock
and takes it into use in bxt-florida audio board driver. The sleep clock
is needed for audio jack plug-in debounce filtering. This also
corrects the clocking in the bxt-florida to avoid FLL locking errors
and use MCLK1 instead of BCLK for audio playback.

Signed-off-by: Jukka Laitinen <jukka.laitinen@intel.com>