intel-iot-refkit update
Pull in the latest intel-iot-refkit submodule. It brings
a lot of layer structure change so we'll update local conf
and bblayer conf version to notify users build/conf/*
settings need update.

In addition, the following changes to the local repository
are made:

* The new version also brings libxml2 CVE-2016-9318 fix that
conflicts with the patch maintained in this repo. Therefore
we drop the local copy of the patch.

* Drop beignet-550x recipe. The new intel-iot-refkit has the
recipe so the no need to maintain it locally.

* Update polkit UID/GID in refos-group/passwd. intel-iot-refkit
drops custom 'base-passwd' settings (it used a subset of
base-passwd groups and passwd). The new settings bring 'sudo'
group that uses GID 27 so it conflicts with that is set for
polkitd. To resolve the conflict, new UID/GID for polkitd
are needed.

* Drop tools-interactive packagegroup in base-image.
The new version of intel-iot-refkit no longer provide the
tools-interactive packagegroup (it was not buildable).
The dropped packages are moved to other features.

* Add packagegroup-java-jdk. The new intel-iot-refkit dropped
packagegroup-java-jdk since it was useless (no meta-java) so
the packagegroup-java-jdk is now maintained in meta-ref-os-iot.

* Update python3 dependencies

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>
Signed-off-by: Tuomas Katila <tuomas.katila@intel.com>