Fix kernel crash in i915
This is a series of three kernel patches. First one fixes a crash, which occurs
when no display is connected and the system is waking up from suspend.

The two other ones enable GuC by default. This actually makes the first one
obsolete, but it is still nice to have in case we need to revert back to
execlists later for some reason.

Signed-off-by: Jukka Laitinen <jukka.laitinen@intel.com>
Reviewed-by: Katila, Tuomas <tuomas.katila@intel.com>