Revert to use old intel_ipu4_pipeline_pm functions for ipu4 driver
It is easier to build the y-cable / multiple camera support on top
of the old intel_ipu4_pipeline_pm functions in the ipu4 drivers. The
current v4l2 versions in the kernel don't support all the needed
pad configurations anyhow, and here we can do the required hacks so
that they are all within the ipu4 driver.

Signed-off-by: Jukka Laitinen <jukka.laitinen@intel.com>